move 2d place utilities to same location
