+++ include/hydra/places/2d_places/ellipsoid_math.h
