+++ src/places/2d_places/place_splitting.cpp
@@ -1,11 +1,12 @@
-#include "hydra/frontend/place_2d_split_logic.h"
+#include "hydra/places/2d_places/place_splitting.h"
 
 #include <kimera_pgmo/mesh_delta.h>
 #include <spark_dsg/dynamic_scene_graph.h>
 #include <spark_dsg/node_attributes.h>
 
-#include "hydra/utils/place_2d_ellipsoid_math.h"
-#include "opencv2/imgproc.hpp"
+#include <opencv2/imgproc.hpp>
+
+#include "hydra/places/2d_places/ellipsoid_math.h"
 
 namespace hydra {
 
@@ -180,16 +181,6 @@ inline void decomposePlace(const PointAdaptor& points,
   }
 }
 
-inline void remapPlace2dConnections(Place2dNodeAttributes& attrs,
-                                    const kimera_pgmo::MeshOffsetInfo& offsets) {
-  kimera_pgmo::MeshOffsetInfo::RemapStats info;
-  auto& connections = attrs.mesh_connections;
-  connections = offsets.remapVertexIndices(connections, &info);
-  attrs.min_mesh_index = info.min_index;
-  attrs.max_mesh_index = info.max_index;
-  attrs.has_active_mesh_indices = !info.all_archived;
-}
-
 }  // namespace
 
 void Place2d::updateIndexBounds() {
@@ -292,26 +283,6 @@ bool shouldAddPlaceConnection(const Place2dNodeAttributes& attrs1,
     return true;
   } else {
     return false;
-  }
-}
-
-void remapPlace2dMesh(Place2dNodeAttributes& attrs,
-                      const kimera_pgmo::MeshOffsetInfo& offsets) {
-  remapPlace2dConnections(attrs, offsets);
-
-  auto& indices = attrs.boundary_connections;
-  kimera_pgmo::MeshOffsetInfo::RemapStats info;
-  indices = offsets.remapVertexIndices(indices, &info);
-
-  const auto prev_boundary = attrs.boundary;
-  attrs.boundary.clear();
-  attrs.boundary.reserve(prev_boundary.size());
-  for (size_t i = 0; i < prev_boundary.size(); ++i) {
-    if (info.deleted_indices.count(i)) {
-      continue;
-    }
-
-    attrs.boundary.push_back(prev_boundary[i]);
   }
 }
