+++ src/utils/CMakeLists.txt
@@ -13,6 +13,5 @@ target_sources(
           ${CMAKE_CURRENT_SOURCE_DIR}/pgmo_glog_sink.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/pgmo_mesh_interface.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/pgmo_mesh_traits.cpp
-          ${CMAKE_CURRENT_SOURCE_DIR}/place_2d_ellipsoid_math.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/timing_utilities.cpp
 )
